Add main and game
