+++ snake/main.cpp
@@ -1,0 +1,13 @@
+#include "game.h"
+
+int main(int argc, char **argv) {
+    Game game;
+
+    game.init();
+
+    game.run();
+
+    game.destroy();
+
+    return 0;
+};
