+++ snake/makefile
@@ -3,8 +3,8 @@ CFLAGS = -std=c++17 -Wall -pedantic
 
 all: snake
 
-snake: set_ticker.o snake.o res.o
-	$(CC) -o snake -g set_ticker.o snake.o res.o $(CFLAGS) -lcurses
+snake: set_ticker.o snake.o res.o game.o main.cpp
+	$(CC) -o snake -g set_ticker.o snake.o res.o game.o main.cpp $(CFLAGS) -lcurses
 
 snake.o: snake.cpp snake.h
 	$(CC) -c -g snake.cpp $(CFLAGS)
@@ -14,6 +14,9 @@ set_ticker.o: set_ticker.c
 
 res.o: res.cpp res.h
 	$(CC) -c -g res.cpp $(CFLAGS)
+
+game.o: game.cpp game.h
+	$(CC) -c -g game.cpp $(CFLAGS)
 
 .PHONY: clean
 clean:
